# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupRequest {
  phoneNumber: string;
  password: string;
}

export interface LoginRequest {
  phoneNumber: string;
  password: string;
}

export interface User {
  id: number;
  phoneNumber: string;
  username?: string | null;
  dateOfBirth?: string | null;
  profilePictureUrl?: string | null;
  createdAt: string;
  lastSeen?: string | null;
}

export interface AuthResponse {
  user: User;
  profileComplete: boolean;
}

export interface VerifyDobRequest {
  phoneNumber: string;
  dateOfBirth: string;
}

export interface VerifyDobResponse {
  resetToken: string;
}

export interface ResetPasswordRequest {
  resetToken: string;
  newPassword: string;
}

export interface UpdateProfileRequest {
  username?: string;
  dateOfBirth?: string;
  profilePictureUrl?: string;
}

export interface UserSearchResults {
  users: User[];
}

export interface Message {
  messageId: number;
  chatId: number;
  senderId: number;
  messageText?: string | null;
  fileUrl?: string | null;
  fileType?: string | null;
  fileName?: string | null;
  timestamp: string;
  sender?: User | null;
}

export interface Chat {
  chatId: number;
  otherUser: User;
  lastMessage?: Message | null;
  createdAt: string;
}

export interface ChatList {
  chats: Chat[];
}

export interface CreateChatRequest {
  otherUserId: number;
}

export interface MessageList {
  messages: Message[];
}

export interface SendMessageRequest {
  messageText?: string | null;
  fileUrl?: string | null;
  fileType?: string | null;
  fileName?: string | null;
}

export interface UploadResponse {
  url: string;
  fileType: string;
  fileName: string;
}

export type SearchUsersParams = {
  phone: string;
};

export type GetMessagesParams = {
  before?: string;
  limit?: number;
};

export type UploadFileBody = {
  file: Blob;
};