# lib/api-zod/src/generated/types/chat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";
import type { User } from "./user";

export interface Chat {
  chatId: number;
  otherUser: User;
  lastMessage?: Message | null;
  createdAt: string;
}