# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface Message {
  messageId: number;
  chatId: number;
  senderId: number;
  messageText?: string | null;
  fileUrl?: string | null;
  fileType?: string | null;
  fileName?: string | null;
  timestamp: string;
  sender?: User | null;
}