# lib/api-zod/src/generated/types/messageList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from "./message";

export interface MessageList {
  messages: Message[];
}