# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  username?: string;
  dateOfBirth?: string;
  profilePictureUrl?: string;
}