# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  url: string;
  fileType: string;
  fileName: string;
}