# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandesam Chat API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  phoneNumber: string;
  username?: string | null;
  dateOfBirth?: string | null;
  profilePictureUrl?: string | null;
  createdAt: string;
  lastSeen?: string | null;
}